consumer/requirements.txt
pyspark-stubs==2.3.0
